pyproject.toml
[project]
name = "everybody.codes"
version = "0.1.0"
description = "https://everybody.codes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.6",
    "prettyprinter>=0.18.0",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "termcolor>=3.1.0",
]

[dependency-groups]
dev = [
    "types-python-dateutil>=2.9.0.20250809",
    "types-requests>=2.32.4.20250809",
]

[tool.isort]
profile = "black"
line_length = 79
force_single_line = true
skip_gitignore = true
include_trailing_comma = true

[tool.mypy]
strict = true
python_executable = "$VIRTUAL_ENV/bin/python3"

[[tool.mypy.overrides]]
module = "prettyprinter.*"
ignore_missing_imports = true

[tool.pylsp-mypy]
enabled = true

[tool.ruff]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = ["COM", "D1", "PLR2004", "S101", "T20"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.vulture]
exclude = [".venv"]
min_confidence = 80
